feat(frontend): combine positions and manual trade into single card
ポジション一覧 (PositionPanel) と手動取引 (ManualTradeCard) を 1 枚の
PositionsAndTradeCard に統合する。建玉を確認しながら追加発注・利確
ができるよう近接配置する。建玉の合計含み損益もヘッダに表示する
ようにし、状況把握を 1 か所で完結させる。

旧 PositionPanel.tsx / ManualTradeCard.tsx は他で参照されていない
ため削除する。
