+++ frontend/src/routes/index.tsx
@@ -3,10 +3,9 @@ import { AppFrame } from "../components/AppFrame";
 import { KpiCard } from "../components/KpiCard";
 import { CandlestickChart } from "../components/CandlestickChart";
 import { IndicatorPanel } from "../components/IndicatorPanel";
-import { PositionPanel } from "../components/PositionPanel";
+import { PositionsAndTradeCard } from "../components/PositionsAndTradeCard";
 import { BotControlCard } from "../components/BotControlCard";
 import { LiveTickerCard } from "../components/LiveTickerCard";
-import { ManualTradeCard } from "../components/ManualTradeCard";
 import { OrderbookPanel } from "../components/OrderbookPanel";
 import { ExecutionQualityCard } from "../components/ExecutionQualityCard";
 import { HaltReasonBadge } from "../components/HaltReasonBadge";
@@ -148,14 +147,14 @@ function Dashboard() {
             ofiLong={indicators?.ofiLong ?? null}
           />
           <ExecutionQualityCard />
-          <ManualTradeCard
+          <PositionsAndTradeCard
             symbolId={symbolId}
+            positions={positions}
             currencyPair={currentSymbol?.currencyPair}
             lotStep={currentSymbol?.baseStepAmount}
             minLot={currentSymbol?.minOrderAmount}
           />
           <IndicatorPanel indicators={indicators} />
-          <PositionPanel positions={positions} />
         </aside>
       </div>
     </AppFrame>
